Checkpoint_5: November 7th Update

+++ app/controllers/wikis_controller.rb
@@ -1,36 +1,48 @@
 class WikisController < ApplicationController
+  include Pundit
+  after_action :verify_authorized, except: :index
+  after_action :verify_policy_scoped, only: :index
+
   def index
-    @wikis = Wiki.all
+    @wikis = policy_scope(Wiki)
+    #authorize @wiki
+  end
+
+  def show
+    @wiki = Wiki.find(params[:id])
+    authorize @wiki
   end
 
   def create
     @wiki = Wiki.new(wiki_params)
     @wiki.user = current_user
+    authorize @wiki
 
     if @wiki.save
       flash[:notice] = "Wiki was saved successfully"
       redirect_to @wiki
     else
-      flash.now[:alert] = "There was an error saving the post. Please try again."
+      flash.now[:alert] = "There was an error saving the Wiki. Please try again."
       render :new
     end
   end
 
-  def show
-    @wiki = Wiki.find(params[:id])
-  end
-
   def new
     @wiki = Wiki.new
+    authorize @wiki
   end
 
   def edit
-    @wiki = Wiki.find(params[:id])
+    @user = current_user
+    @users = User.all
+      @wiki = Wiki.find(params[:id])
+      authorize @wiki
   end
 
   def update
     @wiki = Wiki.find(params[:id])
     @wiki.assign_attributes(wiki_params)
+    authorize @wiki
 
     if @wiki.save
       flash[:notice] = "Wiki was updated successfully."
@@ -43,10 +55,12 @@ class WikisController < ApplicationController
 
   def destroy
     @wiki = Wiki.find(params[:id])
+    authorize @wiki
+    title = @wiki.title
 
     if @wiki.destroy
       flash[:notice] = "\"#{@wiki.title}\" was deleted successfully."
-      redirect_to @wiki 
+      redirect_to wikis_path
     else
       flash.now[:alert] = "There was an error deleting the post."
       render :show
@@ -57,14 +71,5 @@ class WikisController < ApplicationController
 
   def wiki_params
     params.require(:wiki).permit(:title, :body, :private)
-  end
-
-  def authorize_user
-    wiki = Wiki.find(params[:id])
-
-    unless current_user == wiki.user
-      flash[:alert] = "You aren't authorized to do that."
-      redirect_to [wiki]
-    end
   end
 end
